adding file manager

+++ apeiron/commons/Files.py
@@ -2,14 +2,20 @@
 # -*- coding: utf-8 -*-
 import click
 import os
-from .message_type import MessageType
 from pathlib import Path
+from apeiron.commons.message_type import MessageType
+from apeiron.commons.file_version_manager import FileVersionManager
 
 class Files:
 
     @staticmethod
-    def save(file, text):
-        file_path_to_save = Path(os.getcwd()) / file
-        with open(file_path_to_save, "w+") as file:
-            file.write(text)
-        return file_path_to_save
+    def save(text, file_name, dir_path_to_save=Path(os.getcwd())):
+        file_path_to_save = dir_path_to_save / file_name
+        if not FileVersionManager.exist(file_path_to_save):
+            with open(file_path_to_save, "w+") as file:
+                file.write(text)
+            return file_path_to_save
+        else:
+            return None
+
+
